# docs/api.yml
openapi: 3.0.3
info:
  title: 散歩ガチャ
  description: 散歩で新しい発見を提供してくれるAPIです。
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: ローカル環境
...
paths:
...
  /api/image:
    post:
      tags: [image]
      security:
        - jwtAuth: []
      summary: 画像を登録
      operationId: createImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateImagePayload"
      responses:
        "201":
          description: 画像の登録成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageResponse"
        "400":
          $ref: "#/components/responses/InvalidError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/UnexpectedError"
  /api/image/{imageId}:
    delete:
      tags: [image]
      security:
        - jwtAuth: []
      summary: 画像を削除
      operationId: deleteImage
      parameters:
        - $ref: "#/components/parameters/ImageIdParam"
      responses:
        "204":
          description: 画像の削除成功
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/UnexpectedError"
components:
  schemas:
...
    CreateImagePayload:
      type: object
      required:
        - historyId
        - file
      properties:
        historyId:
          type: string
          description: 画像が紐づく散歩履歴のID
          example: "history-001"

        file:
          type: string
          format: binary
          description: |
            画像ファイル本体。対応形式は image/jpeg・image/png・image/webp、
            最大5MB。画像のIDはサーバー側で採番する。
    ImageResponse:
      type: object
      required:
        - imageId
        - historyId
        - imagePath
      properties:
        imageId:
          type: string
          description: 画像のID
          example: "image-001"

        historyId:
          type: string
          description: 画像が紐づく散歩履歴のID
          example: "history-001"

        imagePath:
          type: string
          description: |
            画像ファイルの参照パス。HistoryResponse.imagePaths と同じ形式。
            開発環境では /images 配下の静的パス、本番環境ではオブジェクトストレージのURL。
          example: "/images/history-001/9f2c1a4e-6b3d-4c8a-91f2-0a7e5d3b1c4f.jpg"
  parameters:
...
    ImageIdParam:
      name: imageId
      in: path
      required: true
      description: 画像のID
      schema:
        type: string
  responses:
    UnexpectedError:
      description: 予期しないエラー
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "unexpected_error_occurred"
    InvalidError:
      description: 不正なリクエスト
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "invalid_input"
    UnauthorizedError:
      description: 認証されてないユーザのリクエスト
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "authentication_required"
...
    NotFoundError:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "not_found"
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
      description: NextAuth JWTセッション